Move CheckBoxLabel to WellcomeComponents folder
